updates, added some misc. files that make compiling instrumented versions of the FIFO easier.   might remove them from the repository and clone them sep. in the future, but for now this is the fastest way.

+++ randomstring.tcc
@@ -93,7 +93,7 @@ public:
    }
 
 private:
-   const size_t       length;
+   const int8_t       length;
    char               *source;
    gsl_rng            *r;
    const gsl_rng_type *t;
